backend/requirements.txt
# Backend Python Dependencies - Agent M v0.1.0

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Supabase & Database
supabase==2.0.3
postgrest-py==0.13.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# AI & LLM
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Task Queue (for Phase 4)
celery==5.3.4
redis==5.0.1

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
